visNetInteraction

A shiny Module.

This module let you interact with graph Require visnetwork rendered in shiny to have base id `visNetworkId`

A shiny Module.

Usage

mod_visNetInteraction_ui(id)

mod_visNetInteraction_server(
  id,
  igraph_rct,
  e_ignore = c(),
  v_ignore = c(),
  show_hidden = F,
  domain = getDefaultReactiveDomain()
)

mod_visNetModification_ui(id)

mod_visNetModification_server(
  id,
  igraphObj,
  dev = F,
  hard_delete = T,
  NodeAttrTooltip = T,
  EdgeAttrTooltip = T,
  domain = getDefaultReactiveDomain(),
  visNet_options = NULL,
  layout = "layout_nicely"
)

mod_visNet_server(
  id,
  graph_rct,
  debug = F,
  visNet_options = NULL,
  layout = NULL,
  NodeAttrTooltip = F,
  EdgeAttrTooltip = F
)

Arguments

id, input, output, session

Internal parameters for shiny.

igraph_rct

reactive expression for igraph

e_ignore, v_ignore

a vector of node/edge attributes. This tells the interactive UI to stop parsing specific attribute into searchable UI.

show_hidden

used in conjunction with `e_ignore` and `v_ignore`. If you have an attribute begain/has with '.' This flag will show it when set to TRUE. This is useful against attributes created by tidygraph (`.tidygraph_e_index`) or when you have to create color attributes for visNetwork (`color.border`)

domain

session for

igraphObj

a reactive graph object

NodeAttrTooltip, EdgeAttrTooltip

when these two flag are set to ture attributes will be automatically parsed into tooltips, following formula 'attrName: attr'. There is no good parsing for time series type.

visNet_options

list of option passed to `visSetOptions` this could also be reactive

layout

igraph layout to put in `visNetwork::visIgraphLayout` could be static or reactive.

The easiest way to use both modification server and interaction server. mod_visNet_server links both modification and interaction UI and by doing so make it easy

graph_rct

reactive expression of igraph

debug

this flag will create a text output for tracking change.

id

shiny

Value

reactiveValues $Curent and $Main and more

reactive list

Details

$Current is a reactive igraph Object that every is being modified now $Main is the igraph Object that has been committed and saved In addition it return a set of `reactiveValues` which monitor graph changes and track node that is currently clicked.

list of `Current`, `Main` `click_node` and `click_edge` `Main` for graph that has been committed editing; `Current` for graph on display; Two additional `click_node` and `click_edge` let you track edges and node currently clicked